:refacto: api: Add certifiableBadgeAcquisition
